feat: manyTomany with roles and users

+++ database/migrations/2021_07_13_161015_create_phones_table.php
@@ -13,13 +13,18 @@ class CreatePhonesTable extends Migration
      */
     public function up()
     {
-        Schema::create('phones', function (Blueprint $table) {
-            $table->id();
-            $table->string('phone');
-            $table->bigInteger('user_id');
-            $table->timestamps();
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-        });
+        if(!Schema::hasTable('phones')){
+            Schema::create('phones', function (Blueprint $table) {
+                $table->id();
+                $table->string('phone');
+                $table->bigInteger('user_id');
+                $table->timestamps();
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+            });
+        }
+        
+
+        
     }
 
     /**
